# docs/swagger.yaml
basePath: /
definitions:
...
  models.ErrorResponse:
    properties:
      code:
        example: 400
        type: integer
      details:
        example: provider字段不能为空
        type: string
      error:
        example: true
        type: boolean
      message:
        example: 参数验证失败
        type: string
    type: object
...
  models.MessageResponse:
    properties:
      message:
        example: 操作成功
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  models.SyncHistoryEntry:
    properties:
      bill_period:
        example: 2025-09
        type: string
      completed_at:
        example: "2025-09-11T08:15:29Z"
        type: string
      duration:
        example: 125
        type: integer
      error_count:
        example: 0
        type: integer
      id:
        example: sync_123456
        type: string
      message:
        example: 同步完成
        type: string
      provider:
        example: volcengine
        type: string
      record_count:
        example: 1500
        type: integer
      started_at:
        example: "2025-09-11T08:13:24Z"
        type: string
      status:
        example: completed
        type: string
    type: object
  models.SyncHistoryResponse:
    properties:
      count:
        example: 10
        type: integer
      history:
        items:
          $ref: '#/definitions/models.SyncHistoryEntry'
        type: array
    type: object
  models.SyncStatusResponse:
    properties:
      completed_at:
        type: string
      error_count:
        example: 0
        type: integer
      message:
        example: 同步进行中...
        type: string
      progress:
        example: 75.5
        type: number
      provider:
        example: volcengine
        type: string
      record_count:
        example: 1500
        type: integer
      started_at:
        example: "2025-09-11T08:13:24Z"
        type: string
      status:
        example: running
        type: string
    type: object
  models.SyncTriggerRequest:
    properties:
      bill_period:
        example: 2025-09
        type: string
      force:
        example: false
        type: boolean
      parameters:
        additionalProperties: true
        type: object
      provider:
        example: volcengine
        type: string
    required:
    - provider
    type: object
...
  models.TaskListResponse:
    properties:
      count:
        example: 5
        type: integer
      tasks:
        items:
          $ref: '#/definitions/models.TaskResponse'
        type: array
    type: object
  models.TaskRequest:
    properties:
      description:
        example: 同步火山云账单数据
        type: string
      name:
        example: volcengine_sync_task
        type: string
      parameters:
        additionalProperties: true
        type: object
      provider:
        example: volcengine
        type: string
      schedule:
        example: 0 2 * * *
        type: string
      type:
        example: sync
        type: string
    required:
    - name
    - provider
    - type
    type: object
  models.TaskResponse:
    properties:
      completed_at:
        example: "2025-09-11T08:13:30Z"
        type: string
      created_at:
        example: "2025-09-11T08:13:24Z"
        type: string
      description:
        example: 同步火山云账单数据
        type: string
      id:
        example: task_123456
        type: string
      name:
        example: volcengine_sync_task
        type: string
      parameters:
        additionalProperties: true
        type: object
      provider:
        example: volcengine
        type: string
      schedule:
        example: 0 2 * * *
        type: string
      started_at:
        example: "2025-09-11T08:13:24Z"
        type: string
      status:
        example: running
        type: string
      type:
        example: sync
        type: string
      updated_at:
        example: "2025-09-11T08:13:24Z"
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@goscan.com
    name: API Support
  description: |-
    云账单数据同步服务API文档
    支持火山云、阿里云等多云平台的账单数据自动同步到ClickHouse数据库
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Goscan API
  version: "1.0"
paths:
...
  /sync:
    get:
      consumes:
      - application/json
      description: Returns summary status information of all current data sync tasks,
        including running task count, total task count and other statistics
      produces:
      - application/json
      responses:
        "200":
          description: Sync status retrieved successfully
          schema:
            $ref: '#/definitions/models.SyncStatusResponse'
      summary: Get data sync status
      tags:
      - Data Sync
    post:
      consumes:
      - application/json
      description: |-
        Immediately trigger billing data sync operation for specified cloud provider. Supports multiple cloud providers like Volcengine, Alibaba Cloud, etc.
        Supports configuring sync mode, time range, granularity and other parameters.
      parameters:
      - description: Sync request parameters including cloud provider, sync mode,
          billing period range and other configurations
        in: body
        name: sync
        required: true
        schema:
          $ref: '#/definitions/models.SyncTriggerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Sync task triggered successfully
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "400":
          description: Invalid request parameters or missing cloud provider configuration
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Manually trigger billing data sync
      tags:
      - Data Sync
  /sync/history:
    get:
      consumes:
      - application/json
      description: Returns execution records of historical sync tasks, including successful
        and failed tasks, as well as detailed information of each sync
      parameters:
      - default: 100
        description: Limit of returned records
        in: query
        maximum: 1000
        minimum: 1
        name: limit
        type: integer
      - description: Filter records by specified cloud provider
        enum:
        - volcengine
        - alicloud
        - aws
        - azure
        - gcp
        in: query
        name: provider
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: History records retrieved successfully
          schema:
            $ref: '#/definitions/models.SyncHistoryResponse'
      summary: Get sync history records
      tags:
      - Data Sync
...
  /tasks:
    get:
      consumes:
      - application/json
      description: Returns detailed information of all tasks in the system, including
        running, completed and failed tasks
      produces:
      - application/json
      responses:
        "200":
          description: Task list retrieved successfully
          schema:
            $ref: '#/definitions/models.TaskListResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get all task list
      tags:
      - Task Management
    post:
      consumes:
      - application/json
      description: Create and asynchronously execute a new cloud provider billing
        data sync task. Supports multiple cloud providers like Volcengine, Alibaba
        Cloud, etc.
      parameters:
      - description: Task request parameters including task type, cloud provider,
          sync configuration, etc.
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/models.TaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Task created successfully
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Create new sync task
      tags:
      - Task Management
  /tasks/{id}:
    delete:
      consumes:
      - application/json
      description: Cancel running tasks or delete completed task records. Running
        tasks will be forcibly stopped.
      parameters:
      - description: Unique task identifier ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Task cancelled/deleted successfully
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "400":
          description: Invalid task ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Task not found or cannot be cancelled
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Cancel or delete task
      tags:
      - Task Management
    get:
      consumes:
      - application/json
      description: Get detailed information of a specific task by task ID, including
        task status, execution progress, error information, etc.
      parameters:
      - description: Unique task identifier ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Task details retrieved successfully
          schema:
            $ref: '#/definitions/models.TaskResponse'
        "400":
          description: Invalid task ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get specific task details
      tags:
      - Task Management
schemes:
- http
- https
swagger: "2.0"